% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_age_gender}
\alias{census_age_gender}
\title{census_age_gender}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{age}}{Age broken into age bands}
\item{\code{count}}{The count of people}
}}
\usage{
census_age_gender
}
\description{
Age and gender data from the 2016 Australian census. Data is identified at
the most granular statistical level (SA1) that the Australian Bureau of
Statistics publishes data to.
}
\details{
This data set does include total population counts for Australia.
}
\keyword{datasets}
