% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_birthplace}
\alias{census_birthplace}
\title{census_birthplace}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{birthplace}}{The birthplace of the person}
\item{\code{count}}{The count of people}
}}
\usage{
census_birthplace
}
\description{
The birthplace of the person being counted in the 2016 Australian census.
Data is identified at the most granular statistical level (SA1) that the
Australian Bureau of Statistics publishes data to.
}
\keyword{datasets}
