% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_count_location}
\alias{census_count_location}
\title{census_count_location}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{location}}{The location of the person on counting night}
\item{\code{count}}{The count of people}
}}
\usage{
census_count_location
}
\description{
Location of where people where on counting night of the 2016 Australian
census. Data is identified at the most granular statistical level (SA1) that
the Australian Bureau of Statistics publishes data to.
}
\keyword{datasets}
