% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-data.R
\docType{data}
\name{census_g01_citizen}
\alias{census_g01_citizen}
\title{census_g01_citizen}
\format{A data frame / `data.table`  with three variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{count}}{The count of people}
}}
\usage{
census_g01_citizen
}
\description{
The count of citizens in the 2016 Australian census.
Data is identified at the most granular statistical level (SA1) that the
Australian Bureau of Statistics publishes data to.
}
\keyword{datasets}
