% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01-data.R
\docType{data}
\name{census_g01_highest_schooling}
\alias{census_g01_highest_schooling}
\title{census_g01_highest_schooling}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{highest_school_year}}{Highest year of schooling education}
\item{\code{count}}{The count of people}
}}
\usage{
census_g01_highest_schooling
}
\description{
The highest level of schooling achieved. It does not include tertiary
education. Data is identified at the most granular statistical level (SA1)
that the Australian Bureau of Statistics publishes data to.
}
\details{
This data set does include total population counts for Australia.
}
\keyword{datasets}
