% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g02-data.R
\docType{data}
\name{census_g02_all}
\alias{census_g02_all}
\title{census_g02_all}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{median_age_persons}}{Median age of persons}
\item{\code{median_mortgage_repayment_weekly}}{Median mortgage repayment ($/week)}
\item{\code{median_rent_weekly}}{Median rent ($/week)}
\item{\code{median_tot_personal_inc_weekly}}{Median total personal income ($/week)}
\item{\code{median_tot_family_inc_weekly}}{Median total family income ($/week)}
\item{\code{median_tot_household_inc_weekly}}{Median total household income ($/week)}
\item{\code{average_household_size}}{Average household size (Number of people)}
\item{\code{average_num_persons_per_bedroom}}{Average number of persons per bedroom}
}}
\usage{
census_g02_all
}
\description{
Selected medians and averages relating to age, income, housing expenses and
household size.
}
\keyword{datasets}
