% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g06-data.R
\docType{data}
\name{census_g06_marital_status}
\alias{census_g06_marital_status}
\title{census_g06_marital_status}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{age}}{Age broken into age bands}
\item{\code{marital_status}}{Social marital status}
\item{\code{count}}{The count of people}
}}
\usage{
census_g06_marital_status
}
\description{
The social marital status for Australians broken down by age and gender
}
\keyword{datasets}
