% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g01_data.R
\docType{data}
\name{census_indigenous}
\alias{census_indigenous}
\title{census_indigenous}
\format{A data frame / `data.table`  with four variables:
\describe{
\item{\code{SA1_7DIGITCODE_2016}}{A 7 digit identifier for the Statistical Area 1 (SA1)}
\item{\code{gender}}{Gender}
\item{\code{heritage}}{The indigenous heritage of the Australian it includes
  the values of \code{Aboriginal}, \code{Torres Strait Islander},
  {\code{Aboriginal & Torres Strait Islander} and \code{Total}}}
\item{\code{count}}{The count of people}
}}
\usage{
census_indigenous
}
\description{
The counts of people who identify as indigenous and their heritage. Data is
identified at the most granular statistical level (SA1) that the Australian
Bureau of Statistics publishes data to.
}
\keyword{datasets}
